done with CRUD for products and starting some user management features

+++ database/migrations/2023_10_27_061358_create_products_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->foreignIdFor(User::class);
             $table->string("name");
             $table->text("description")->nullable();
-            $table->integer("selling_price");
+            $table->float("selling_price", 10);
             $table->timestamps();
         });
     }
